% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.set.k.R
\name{calc.k.vec}
\alias{calc.k.vec}
\title{use a holosim spatiotemporal surface to set K in pops}
\usage{
calc.k.vec(N, hsl, gen, years = 21000)
}
\arguments{
\item{N}{base level Ne that gets multiplied by per cell suit.}

\item{hsl}{landscape object used in holosimCell}
}
\description{
use a holosim spatiotemporal surface to set K in pops
}
