% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amova.R
\name{landscape.amova.locus}
\alias{landscape.amova.locus}
\title{uses functions in ade4 to calcuate phi-st for a particular locus}
\usage{
landscape.amova.locus(l = 1, rland)
}
\arguments{
\item{l}{locus number}

\item{rland}{landscape object}
}
\value{
list of amova results for a locus
}
\description{
Runs an amova on a locus.  Does not include information about sequence
similarity or ssr size in analysis.
}
\details{
Should be the same as Weir and Cockerham's theta
}
\seealso{
landscape.amova, landscape.amova.pairwise
}
\keyword{misc}
