% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amova.R
\name{landscape.amova.pairwise}
\alias{landscape.amova.pairwise}
\title{calculates pairwise phi_ST for a landscape}
\usage{
landscape.amova.pairwise(rland)
}
\arguments{
\item{rland}{landscape object}
}
\description{
pairwise phi_ST calculator.  Kind of slow. use \code{landscape.sample} to
reduce the size of the calculation.
}
\seealso{
landscape.amova, landscape.amova.locus
}
