% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amova.R
\name{landscape.amova}
\alias{landscape.amova}
\title{calcuates phi-st for every locus in the landscape}
\usage{
landscape.amova(rland, np = 24, ns = 24)
}
\arguments{
\item{rland}{landscape object}

\item{np}{max number of pops to include}

\item{ns}{max number of samples to collect}
}
\value{
vector of length equal to the number of loci
}
\description{
calcuates phi_ST for every locus in the landscape
}
\seealso{
\code{\link{landscape.amova.locus}},\code{\link{landscape.amova.pairwise}}
}
\keyword{misc}
