% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepopR.R
\name{landscape.genepop.output}
\alias{landscape.genepop.output}
\title{writes genepop files based on the genotypes of individuals in a landscape}
\usage{
landscape.genepop.output(
  l,
  fn = "genepop.out",
  title = "rmetasim landscape output"
)
}
\arguments{
\item{l}{landscape object}

\item{fn}{output file name}

\item{title}{title of output in genepop file}
}
\description{
creates a genepop file.  Occupied habitats correspond to genepop
populations.
}
\details{
Only exports diploid data.  You might want to run \code{landscape.sample()}
on the landscape before exporting it
}
\examples{

#	l <- landscape.new.example()	
#	l <- landscape.simulate(l,5)
#	landscape.genepop.output(landscape.sample(l,3,24))

}
\seealso{
landscape.write.foreign, landscape.write.fdist
}
\keyword{misc}
