% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_stats.R
\name{landscape.hab.stats}
\alias{landscape.hab.stats}
\title{calculate summary statistics on habitats in 2d landscape}
\usage{
landscape.hab.stats(l, e = 1)
}
\arguments{
\item{l}{quantsel landscape}

\item{e}{epoch (default 1, usually fine)}
}
\value{
a named list of cell stats
}
\description{
calculate summary statistics on habitats in 2d landscape
}
