\name{landscape.new.expression}
\alias{landscape.new.expression}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{don't use}
\description{
 not implemented 
}
\usage{
landscape.new.expression(rland, expmat, hsq)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rland}{landscape object}
  \item{expmat}{expression matrix}
  \item{hsq}{narrow sense heritabilities}
}
\details{
This is only skeleton code.  
}
\value{
Not sure what it will return
}
\author{Allan Strand}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}% __ONLY ONE__ keyword per line
