% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.nophen.R
\name{landscape.nophen}
\alias{landscape.nophen}
\title{creates a dummy set of quantgen params to allow marker-only simulation (nphen has to be 0)}
\usage{
landscape.nophen(rland)
}
\arguments{
\item{rland}{a quantsel landscape object.  Assumes no phneotypes}
}
\description{
creates a dummy set of quantgen params to allow marker-only simulation (nphen has to be 0)
}
