\name{landscape.test.function}
\alias{landscape.test.function}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ tests whatever c code is compiled in}
\description{
 currently generates 10,000 random pulls from a pdf
}
\usage{
landscape.test.function()
}

\value{
vector
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}

