% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdist.R
\name{landscape.write.fdist}
\alias{landscape.write.fdist}
\title{writes a landscapes allele frequencies to an FDIST file}
\usage{
landscape.write.fdist(l, popnum = 20, samp = 24)
}
\arguments{
\item{l}{landscape object}

\item{popnum}{max number of populations to sample}

\item{samp}{max number of individuals to sample per population}
}
\description{
Calculates the allele frequencies in each occupied habitat up to the number
specified in popnum.  Makes calculations based upon the number of
individuals sampled stored in samp
}
\seealso{
\code{\link{landscape.write.foreign}},\code{\link{landscape.genepop.output}}
}
\keyword{misc}
