% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2popcrds.R
\name{raster2popcrds}
\alias{raster2popcrds}
\title{Extract resolution and extents to convert a raster to quantsel popcoords}
\usage{
raster2popcrds(r, zeroOrigin = TRUE)
}
\arguments{
\item{r}{raster input}

\item{zeroOrigin}{make sure that no coordinates were negative}
}
\description{
Extract resolution and extents to convert a raster to quantsel popcoords
}
