\name{landscape.amova.pairwise}
\alias{landscape.amova.pairwise}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ calculates pairwise phi_ST for a landscape}
\description{
  pairwise phi_ST calculator.  Kind of slow. use \code{landscape.sample}
  to reduce the size of the calculation.
}
\usage{
landscape.amova.pairwise(rland)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rland}{landscape object}
}

\seealso{ landscape.amova, landscape.amova.locus}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.


